pyproject.toml
[project]
name = "app-save-money"
version = "0.1.0"
dependencies = [
    "flet>=0.21.0",
    "google-generativeai",
    "python-dotenv",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = ["E501"] # Exemplo de ignore

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
